+++ src/providers/emoji/Emojis.cpp
@@ -196,7 +196,7 @@ void Emojis::loadEmojis()
 
         this->emojiFirstByte_[emojiData->value.at(0)].append(emojiData);
 
-        this->emojis.insert(emojiData->unifiedCode, emojiData);
+        this->emojis.push_back(emojiData);
 
         if (unparsedEmoji.HasMember("skin_variations"))
         {
@@ -218,8 +218,7 @@ void Emojis::loadEmojis()
                 this->emojiFirstByte_[variationEmojiData->value.at(0)].append(
                     variationEmojiData);
 
-                this->emojis.insert(variationEmojiData->unifiedCode,
-                                    variationEmojiData);
+                this->emojis.push_back(variationEmojiData);
             }
         }
     }
@@ -244,10 +243,8 @@ void Emojis::sortEmojis()
 void Emojis::loadEmojiSet()
 {
     getSettings()->emojiSet.connect([this](const auto &emojiSet) {
-        this->emojis.each([=](const auto &name,
-                              std::shared_ptr<EmojiData> &emoji) {
-            (void)name;
-
+        for (const auto &emoji : this->emojis)
+        {
             QString emojiSetToUse = emojiSet;
             // clang-format off
             static std::map<QString, QString> emojiSets = {
@@ -289,7 +286,7 @@ void Emojis::loadEmojiSet()
             emoji->emote = std::make_shared<Emote>(Emote{
                 EmoteName{emoji->value}, ImageSet{Image::fromUrl({url}, 0.35)},
                 Tooltip{":" + emoji->shortCodes[0] + ":<br/>Emoji"}, Url{}});
-        });
+        }
     });
 }
 
@@ -430,7 +427,7 @@ QString Emojis::replaceShortCodes(const QString &text) const
     return ret;
 }
 
-const EmojiMap &Emojis::getEmojis() const
+const std::vector<EmojiPtr> &Emojis::getEmojis() const
 {
     return this->emojis;
 }